requirements.txt
win32gui
numpy==1.15.3
matplotlib==3.0.1
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
pillow==8.4.0
pypiwin32==223
pywin32==303
mouse==0.7.1